Fix v1.0F raw delete global id type

+++ scripts/apply_ghostbase_v10f_raw_difference_snapshot_probe.py
@@ -145,7 +145,7 @@ private func ghostBaseV10FRawDeleteHit(messageIds: [MessageId], source: String) 
     }
 }
 
-private func ghostBaseV10FRawDeleteGlobalHit(globalIds: [Int64], source: String) {
+private func ghostBaseV10FRawDeleteGlobalHit(globalIds: [Int32], source: String) {
     ghostBaseV10FRawRecord("deleteGlobalSeen")
     ghostBaseV10FRawRecord("deleteGlobalIds", amount: globalIds.count)
     ghostBaseV10FRawSet("LastDeleteSource", source)
